mcps/glpi/pyproject.toml
[project]
name = "glpi-mcp-server"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.0.2",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
]
